comments: refactor code for handling cut/paste properly
