+++ src/packages/frontend/frame-editors/code-editor/codemirror-editor.tsx
@@ -36,6 +36,7 @@ import { Actions } from "./actions";
 import { EditorState } from "../frame-tree/types";
 import { Path } from "../frame-tree/path";
 import { initFold, saveFold } from "@cocalc/frontend/codemirror/util";
+import { initCommentCutPasteSupport } from "@cocalc/frontend/frame-editors/generic/comments/codemirror";
 
 const STYLE = {
   width: "100%",
@@ -326,75 +327,12 @@ export const CodemirrorEditor: React.FC<Props> = React.memo((props) => {
       SAVE_DEBOUNCE_MS,
     );
 
-    let cutCopyBuffer: { [text: string]: any[] } = {};
+    initCommentCutPasteSupport(cm);
 
     cm.on("beforeChange", (_, changeObj) => {
       if (changeObj.origin == "paste") {
         // See https://github.com/sagemathinc/cocalc/issues/5110
         save_syncstring();
-      } else if (changeObj.origin == "cut") {
-        const text = cm.getRange(changeObj.from, changeObj.to);
-        console.log("cut", changeObj);
-        const marks: any[] = [];
-        const lines = cm.getValue().split("\n");
-        let start = positionToLinear(lines, changeObj.from);
-        for (const mark of cm.getAllMarks()) {
-          const loc = mark.find() as null | CodeMirror.MarkerRange;
-          if (loc == null) {
-            continue;
-          }
-          if (isInRange(loc.from, changeObj) && isInRange(loc.to, changeObj)) {
-            const from = positionToLinear(lines, loc.from) - start;
-            const to = positionToLinear(lines, loc.to) - start;
-            marks.push({ from, to, mark });
-          }
-        }
-        if (marks.length > 0) {
-          cutCopyBuffer[text] = marks;
-        }
-        console.log({ cutCopyBuffer });
-        editor_actions()?.comments.saveComments();
-      }
-    });
-
-    cm.on("changes", (_, v) => {
-      let pasted = false;
-      let k = 0;
-      for (const changeObj of v) {
-        if (changeObj.origin == "paste") {
-          pasted = true;
-          if (cutCopyBuffer != null) {
-            console.log({
-              a: changeObj.text.join("\n"),
-              cutCopyBuffer,
-            });
-            const marks = cutCopyBuffer[changeObj.text.join("\n")];
-            console.log("paste", { marks });
-            if (marks != null) {
-              const lines = cm.getValue().split("\n");
-              const start = k + positionToLinear(lines, changeObj.from);
-              for (const { from, to, mark } of marks) {
-                const cur = mark.find();
-                if (cur != null) {
-                  // it didn't get removed
-                  continue;
-                }
-                const loc = {
-                  from: linearToPosition(lines, from + start),
-                  to: linearToPosition(lines, to + start),
-                };
-                setMarkLocation({ doc: cm.getDoc(), loc, mark });
-              }
-            }
-            k += changeObj.text.join("\n").length;
-          }
-        }
-      }
-      console.log("after paste", pasted);
-      if (pasted) {
-        for (const prop in cutCopyBuffer) {
-          delete cutCopyBuffer[prop];
-        }
       }
     });
 
@@ -586,44 +524,4 @@ if ((CodeMirror as any).commands.save == null) {
       f(true);
     }
   };
-}
-
-import { setMarkLocation } from "@cocalc/frontend/frame-editors/generic/comments/comments";
-import { cmp } from "@cocalc/util/misc";
-function cmp_pos(a: CodeMirror.Position, b: CodeMirror.Position) {
-  const c = cmp(a.line, b.line);
-  if (c) {
-    return c;
-  }
-  return cmp(a.ch, b.ch);
-}
-
-function isInRange(
-  a: CodeMirror.Position,
-  range: { from: CodeMirror.Position; to: CodeMirror.Position },
-): boolean {
-  return cmp_pos(range.from, a) <= 0 && cmp_pos(a, range.to) <= 0;
-}
-
-function positionToLinear(lines: string[], pos: CodeMirror.Position): number {
-  if (pos.line >= lines.length) {
-    throw Error("invalid position");
-  }
-  let n = 0;
-  for (let i = 0; i < pos.line; i++) {
-    n += lines[i].length + 1;
-  }
-  return n + pos.ch;
-}
-
-function linearToPosition(lines: string[], n: number): CodeMirror.Position {
-  for (let line = 0; line < lines.length; line++) {
-    const m = lines[line].length;
-    if (n > m) {
-      n -= m + 1;
-    } else {
-      return { line, ch: n };
-    }
-  }
-  return { line: lines.length, ch: 0 };
 }
